Refactor pretty printing
